Extract Index component
