+++ lib/annotate_rb/model_annotator/index_annotation.rb
@@ -4,6 +4,7 @@ module AnnotateRb
   module ModelAnnotator
     module IndexAnnotation
       autoload :AnnotationBuilder, "annotate_rb/model_annotator/index_annotation/annotation_builder"
+      autoload :IndexComponent, "annotate_rb/model_annotator/index_annotation/index_component"
     end
   end
 end
